+++ src/navigation/MainTabNavigator.js
@@ -4,7 +4,7 @@ import { createStackNavigator, createBottomTabNavigator } from 'react-navigation
 
 import TabBarIcon from '../components/TabBarIcon';
 import HomeScreen from '../screens/HomeScreen';
-import LinksScreen from '../screens/LinksScreen';
+import InstructionsScreen from '../screens/InstructionsScreen';
 import SettingsScreen from '../screens/SettingsScreen';
 import LoginView from '../screens/LoginView';
 
@@ -27,11 +27,11 @@ HomeStack.navigationOptions = {
 };
 
 const LinksStack = createStackNavigator({
-  Links: LinksScreen,
+  Links: InstructionsScreen,
 });
 
 LinksStack.navigationOptions = {
-  tabBarLabel: 'Links',
+  tabBarLabel: 'Instructions',
   tabBarIcon: ({ focused }) => (
     <TabBarIcon
       focused={focused}
